TIP-221: Move AttributeValuesResolver
