+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -9,7 +9,7 @@ parameters:
     pim_catalog.manager.product_template_media.class:    Pim\Bundle\CatalogBundle\Manager\ProductTemplateMediaManager
     pim_catalog.applier.product_template.class:          Pim\Bundle\CatalogBundle\Manager\ProductTemplateApplier
     pim_catalog.resolver.variant_group_attributes.class: Pim\Bundle\CatalogBundle\Manager\VariantGroupAttributesResolver
-    pim_catalog.resolver.attribute_values.class:         Pim\Bundle\CatalogBundle\Manager\AttributeValuesResolver
+    pim_catalog.resolver.attribute_values.class:         Pim\Component\Catalog\Manager\AttributeValuesResolver
 
 services:
     # Managers
